% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optfwhm.R
\name{optfwhm}
\alias{optfwhm}
\title{Optimal smoothing parameter for All-resolutions inference cluster thresholding in neuroimaging (opt_fwhm)}
\usage{
optfwhm(zstats, min_fwhm = 2, max_fwhm = 12, tdp, max_size = T)
}
\arguments{
\item{min_fwhm}{Minimum level of spatial smoothing a researcher wishes to set, default is 2.}

\item{max_fwhm}{Maximum level of spatial smoothing a researcher wishes to set, default is 2.}

\item{tdp}{Minimum TDP threshold set to apply All-Resolutions Inference framework}

\item{max_size}{Optimization based on maximal size (True) or maximal number of clusters (False)}

\item{zstat}{Z-stats 3D array of activation values, or a nifti file name.}
}
\value{
Returns a value that can be used to define the optimal smoothing parameter.
}
\description{
The function is made to guide researchers in defining their optimal smoothing parameter
}
\examples{

optfwhm(zstat, min_fwhm = 2, max_fwhm = 12, tdp = 0.7, max_size = T)





}
\author{
Jonathan Ornstein, Wouter Weeda.
}
